[native] Introduce RegistrationMode in EditUserAvatarProvider
Summary:
A consumer of the `EditUserAvatarContext` will be able to call `setRegistrationMode` to override what happens when an avatar is successfully selected.

This will allow the registration flow to get the avatar info without needing to change much other code.

Test Plan:
This was tested in combination with the rest of the stack. I tested:

1. ENS flow, both when user has an ENS avatar and when they don't
2. Setting an emoji avatar
3. Setting an image avatar via camera
4. Setting an image avatar via media library
5. Tested avatar update flow after being logged in to make sure I didn't regress anything

Reviewers: ginsu, atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D8041