[native] Introduce new EmojiAvatarSelection screen for registration flow
Summary:
For emoji avatar selection, our current pattern in the `native` app is to show another screen that handles this.

This diff continues that pattern, but introduces a new screen for the registration flow. Luckily we already have the scaffolding for this, as we needed separate screens for `EmojiThreadAvatarCreation` and `EmojiUserAvatarCreation` already.

Depends on D8042

Test Plan:
This was tested in combination with the rest of the stack. I tested:

1. ENS flow, both when user has an ENS avatar and when they don't
2. Setting an emoji avatar
3. Setting an image avatar via camera
4. Setting an image avatar via media library
5. Tested avatar update flow after being logged in to make sure I didn't regress anything

Reviewers: ginsu, atul

Subscribers: