[native] Make EditUserAvatar component conditionally navigate to registration screens
Summary:
For the registration flow, we have to render different screens for camera and emoji selection.

This diff makes it possible to use `EditUserAvatar` in both contexts, by conditionally deciding which screen to navigate to based on `EditUserAvatarContext` state.

Depends on D8046

Test Plan:
This was tested in combination with the rest of the stack. I tested:

1. ENS flow, both when user has an ENS avatar and when they don't
2. Setting an emoji avatar
3. Setting an image avatar via camera
4. Setting an image avatar via media library
5. Tested avatar update flow after being logged in to make sure I didn't regress anything

Reviewers: ginsu, atul

Reviewed By: atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D8047