[web][fix] Fixed cursor issue in the message editing
Summary:
There was a bug, where after typing the character the cursor was moved to the end of the text area. It made editing the middle of the content very cumbersome. This address this issue, now `focusAndUpdateText()` method is called only once, not after every text change.

Raised in https://linear.app/comm/issue/ENG-3981/[web]-fix-cursor-moving-to-the-end-of-the-textbox-after-character

Test Plan:
- Entered edit mode on the web
- Checked if the textarea is focused
- Edited the message in the middle, checked if the cursor is not moved to the end

Reviewers: michal, inka, ashoat

Reviewed By: ashoat

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D8035