[keyserver] Add fetch latest messages endpoint
Summary:
[ENG-3972](https://linear.app/comm/issue/ENG-3972)
Part of the fix for the "no messages" bug. This endpoint allows us to fetch the latest message for 25 threads, that we still don't have messages for (as an argument you pass the oldest message the client is aware of).

This is mostly based on @ashoat idea in the parent task

Test Plan:
Tested with the later diffs if the messages are fetched correctly. Tests manully by making requests and checking if correct messages were returned.

Reviewers: inka, kamil, kuba

Subscribers: