[landing] introduce request access compoonent
Summary: As part of the redesign for the landing page there is Request access call to action at the bottom of the main page. This diff creates that section into its own component. Here is a screenshot of the designs for extra context

Test Plan: Please watch the demo video below/see the screenshots

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8080