[native] Default to ENS avatar for users that have one
Summary:
In the new registration flow, I want to default to using an ENS avatar if the user has one.

Depends on D8048

Test Plan:
This was tested in combination with the rest of the stack. I tested:

1. ENS flow, both when user has an ENS avatar and when they don't
2. Setting an emoji avatar
3. Setting an image avatar via camera
4. Setting an image avatar via media library
5. Tested avatar update flow after being logged in to make sure I didn't regress anything

Since this is the last diff in the stack, I'll also include a video: {F566716}

Reviewers: atul, ginsu

Reviewed By: atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D8049