[web] Preload encrypted media
Summary:

Resolves [[ https://linear.app/comm/issue/ENG-3679/preload-encrypted-media-on-web | ENG-3679 ]].
Used `fetch` to preload encrypted media. The current `preloadImage` function used to preload them by creating a dummy `<img />` element and setting the `src` property, which was not suitable for encrypted media.

The introduced function in fact preloads any request, including images, but I left the original because browser might have some optimizations for images. Also it is used in other places in the codebase e.g. to get the dimensions.

Test Plan:

Entered an encrypted media chat and selected a photo to send. Observed Dev Tools Network tab to see that the encrypted media was preloaded:

Reviewers: atul, ashoat

Subscribers: