[web] Change border radius of filled-in input fields
Summary:
issue: https://linear.app/comm/issue/ENG-3866/update-web-filled-in-input-fields
We want all filled in input fields to have a border radius of 8.

Test Plan:
Checked that the radius of search input fields and the input bar is now 8px.
I went through all places in the code where border radius is specified and checked what they relate to. I went through the app and checked for any filled-in input fields.

Reviewers: kuba, michal, ginsu

Reviewed By: michal

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7954