[native] Change border radius of filled-in input fields
Summary:
issue: https://linear.app/comm/issue/ENG-3867/update-mobile-filled-in-input-fields
We want all filled in input fields to have a border radius of 8.

Test Plan:
Checked that the radius of the input bar is now 8px.
I went through the app and checked for any filled-in input fields that don't have a border radius of 8. I went throught the code and checked all places using TextInput (and code using that code if
the style was not specified there, and so on).

Reviewers: kuba, michal, ginsu

Reviewed By: michal

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7955