[web] Fetch latest messages in chat list
Summary:
Part of [ENG-3976](https://linear.app/comm/issue/ENG-3976)
Fetches more messages for threads, when the user is near the end of the thread list.
Additionaly if there not enough threads to fill the whole thread list, `onItemsRendered` doesn't fire (quirk of `react-window`) so we wouldn't fetch more messages. Because of that there's an `useEffect` that calculates if the current thread list is smaller than the total height and if yes tries fetching more messages.

Depends on D8234

Test Plan:
Try scrolling the thread list and check if threads are correctly loading and there are no jumping threads.

Reviewers: inka, kuba, kamil

Subscribers: