[web] Escape focus trap when an outside node is clicked
Summary: All of the context and investigation is in [[ https://linear.app/comm/issue/ENG-3954/cant-unpin-message-from-pinned-message-modal-on-web | ENG-3954 ]], but encompassing `ModalOverlay` in a focus-trap
was blocking all tooltip actions, since it was not contained within the trapped DOM element.

There were a few solutions proposed, this one is just the simplest. There are two [[ https://github.com/focus-trap/focus-trap#createoptions | configurable options ]] that can be passed into the component,
clickOutsideDeactivates and allowOutsideClick. Both seem to resolve the issue, but I opted for `allowOutsideClick` as instead of entirely deactivating the focus-trap and letting outside clicks to continue, it
just filters through the click and allows outside actions.

Test Plan: Please see the video below showing the solution

Reviewers: kuba!, ashoat

Subscribers: