[identity] register_reserved_password_user_start method
Summary:
This method is similar to the existing register_password_user_start method. The difference is, here we check if the username is in the list of reserved usernames from Ashoat's keyserver, and if it is, we expect a signed message from Ashoat's keyserver attesting that the caller can claim this username.

This RPC should only be called if login fails because the username is already registered on Ashoat's keyserver but not yet registered on the identity service.

Depends on D8061

Test Plan: Seeded the DynamoDB reserved usernames table with a reserved username, then successfully claimed and registered the username with the new RPC and the registration finish RPC

Also tested failure cases where the signed message is malformed, the timestamp expired, and username not actually reserved

Reviewers: jon, bartek

Subscribers: