[landing] introduce blur effect for hero image
Summary: When I was rebuilding the new hero section, I decided to remove the blur effect from the hero image. I did this for a lot of reasons, but mainly removing the effect from the image asset made the image a lot more clear. I think the blur effect is still great, so I readded it back in as a div element. Please see the designs for context

Test Plan: Please look at the demo videos/screenshots below

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8076