[landing] introduce glow effect for hero image
Summary:
When I was rebuilding the new hero section, I decided to remove the glow effect from the hero image. I did this for a lot of reasons, but mainly removing this effect from the image asset made the image a lot more clear. I think the glow effect is still great, so I added it back in as a div element. Please see the designs for context

{F569662}

{F569661}

Depends on D8075

Test Plan:
Please look at the demo videos/screenshots below

Desktop/Responsive:
{F569663}

Tablet:
{F569669}

Mobile:
{F569670}

Reviewers: atul, kamil

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8076