[landing] move hero section code into own component
Summary:
Simple move diff. All other sections in `AppLanding` will be put in their own component so I figured it would be good to have one for the hero as well.

Depends on D8076

Test Plan: Double checked that the hero section still looked and behaved as it did previously

Reviewers: atul, kamil

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8077