[landing] repurpose InfoBlock component
Summary:
This diff repurposes the `InfoBlock` component to work for this landing page redesign. A big change with this `InfoBlock` component is that we no longer need an image, and all the content in the `InfoBlock` is centered. Here is a screenshot of the designs for extra context:

{F569825}

{F569824}

Depends on D8078

Test Plan:
Please see the demo videos/screenshots below to see how the new `InfoBlock` component:

Desktop/Responsive:
{F569811}

Tablet:
{F569826}

Mobile:
{F569827}
{F569832}

Reviewers: atul, kamil

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8079