[landing] repurpose InfoBlock component
Summary: This diff repurposes the `InfoBlock` component to work for this landing page redesign. A big change with this `InfoBlock` component is that we no longer need an image, and all the content in the `InfoBlock` is centered. Here is a screenshot of the designs for extra context

Test Plan: Please see the demo videos/screenshots below to see how the new `InfoBlock` component

Reviewers:

Subscribers: