Implement schema for new JSI module - CommValidationModule
Summary: This differential introduces new JSI module to the project - CommValidationModule. Its intention is to enforce in JS things we assume about JS in C++ and vice versa. Currently it contains only one
method - validateMessageTypes, which aim is to encure all MessageSpecs implemented in JS are implemented in C++ as well.

Test Plan: bartek

Reviewers:

Subscribers: