[native] Introduce `CommunityCreationContentContainer` component
Summary:
This component was heavily inspired by `RegistrationContentContainer`.

Replaced the `ScrollView` in `RegistrationContentContainer` with a `View` to avoid issue with virtualized list being nested within `ScrollView`:

{F568067}

Kept all the other styling and whatnot for consistency.

As with `RegistrationContentContainer`, this component can probably be made more general, "added to the design system," and moved to `/components` in the future. Will discuss how to approach this w/ @ted when he's back.

---

Depends on D8056

Test Plan: Things look as they did before and the RedBox no longer appears when the screen renders.

Reviewers: ashoat, ginsu

Reviewed By: ashoat

Subscribers: tomek, ted

Differential Revision: https://phab.comm.dev/D8057