[web] Create a model that allows admins to change a member's role (dropdown menu, actionable buttons)
Summary: This is the second part of the modal creation, utilizing the earlier `Dropdown` component and creating the two 'back' and 'save' buttons. Functionality for these buttons will be added up the stack to
make this diff easier to review.

https://linear.app/comm/issue/ENG-4038/create-a-modal-that-allows-admins-to-change-a-members-role

Depends on D8137

Test Plan: Please see a screenshot of what the complete modal looks like now

Reviewers: atul, ginsu

Subscribers: