[native] Fix userInfo calculation in UserAvatar
Summary:
The previous code was passing the wrong `userInfo` to `useENSResolvedAvatar` when the `userInfo` prop was specified. The new code keeps the decision in one place, with only one selector.

Depends on D8086

Test Plan: Make sure we're not randomly using the anonymous user emoji from `EditUserAvatar` rendered with a `userInfo` prop

Reviewers: atul, ginsu

Reviewed By: atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D8087