[landing] introduce keyserver faq component
Summary:
This diff is the first part to introduce the keyserver faq component. The focus of this diff is to get the general structure/behavior of the keyserver faq. Subsequent diffs will handle other things like the animation and implementation of this component

Depends on D8092

Test Plan:
Please see the demo video/screenshots below

Desktop/Responsive:
{F572215}

Tablet:
{F572222}

Mobile:
{F572223}

Reviewers: atul, kamil

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8093