[landing] animate the keyserver faq section
Summary:
This diff is the second part to the introduction of the keyserver faq section. The focus of this diff is animating the answers to slide in with an accordion like animation. When I first built this, I ran into the issue where I couldn't add a transition with `auto`. Since we need an explicit value to to make this transition, I initially thought the best way to go about animating this was to calculate the height using JS. However, after further research I found a much more elegant way to achieve the same effect using grid

Sources:
https://developer.mozilla.org/en-US/docs/Web/CSS/CSS_Transitions/Using_CSS_transitions#specifications
https://carlanderson.xyz/how-to-animate-on-height-auto/
https://keithjgrant.com/posts/2023/04/transitioning-to-height-auto/
https://www.youtube.com/watch?v=B_n4YONte5A

Depends on D8093

Test Plan:
Please see the demo video below:

{F574086}

Reviewers: atul, kamil

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8095