[native] Introduce RegistrationContext
Summary:
Explanation for why we need a context is in [ENG-4047](https://linear.app/comm/issue/ENG-4047/code-to-create-an-account-for-new-registration-flow) and [ENG-4043](https://linear.app/comm/issue/ENG-4043/move-react-state-for-registration-flow-to-context).

This diff just introduces an empty context; later diffs will actually implement functionality.

Test Plan: The whole stack was tested as follows:

1. Try creating an account with the default avatar
2. Try creating an account with an emoji avatar
3. Try creating an account with an image avatar (that needs to be uploaded)

I made sure of the following:

1. Loading spinner appeared on the button, not on the avatar, and was present throughout
2. After the `RegistrationNavigator` is dismissed, the avatar that appears for the user's `PRIVATE` chat in the `ChatThreadList` is correct
3. Subsequent delete-account-and-then-register-again worked correctly
4. Various error cases are triggered correctly

Reviewers: atul, ginsu

Subscribers: