[web] Display pinned messages not present in Redux
Summary: As mentioned in [[
https://linear.app/comm/issue/ENG-3924/pinned-messages-not-showing |
ENG-3924 ]], there is an issue with some pinned messages where if they are
not present in Redux, they will not show up in the modal.

This is because `messageListData` will incorrectly set the messageInfo's
`isPinned` flag to false, and therefore we will filter it out. This diff
resolves this on `web` by keeping a Set of pinned messgae IDs, and
filtering out messages that are not part of the set, rather than based on
their `isPinned` value. This maintains the same behaviour for pinned
messages that were fetched, but also covers the case where they are not
part of Redux.

Test Plan: Please see the before and after for a pinned message that had
no reactions/related messages and was not present in Redux. Confirmed
separately that pinned messages on `web` still behave as expected.

Reviewers: atul, inka!, ginsu

Subscribers: