Implement message typoes validator and import it in root.react.js for side effect
Summary:
This differential itroduces a file with custom code that checks if all message specs that might generate notifications are implemented in C++ as well. In the case of missing C++ message spec an dalert is displayed. Additionally this
feature required us to alter MessageSpec type a little bit so that `generatesNotifs` field is now optional.

Test Plan: Build the app and launch it. Ensure alert is not displayed. Go to `message-types-validator.js` file and comment the check for `generatesNotifs` in message spec. Now an alert with 5 message types should be displayed.

Reviewers: tomek, atul, bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D8103