Provided detailed error messages on olm functions failures in CryptoModule
Summary: This differential refactors CryptoModule.cpp and Session.cpp so that each time olm function fails detailed and informative error message is provided.

Test Plan:
It would be cumbersome to attempt to fail each olm function that is used in those files, however the fact that we already used this approach a couple of times:
https://linear.app/comm/issue/ENG-3982/recreate-keyserver-olm-accounts-on-dev-and-production-environments or here: https://linear.app/comm/issue/ENG-3844/unable-to-correctly-instantiate-olm-session-between-the-device-and-the should be enough to
consider it correct.

Reviewers: tomek, bartek, atul

Reviewed By: bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D8114