[kayserver] Fix search endpoint failing when called with a stopword
Summary:
issue: https://linear.app/comm/issue/ENG-4055/search-endpoint-failing-when-called-with-only-a-stopword
When `searchMessagesInSingleChat` was passed a string with ononly stopwords, they were stripped away by `processQueryForSearch` and the `pattern` variable was `'+'`, which is not a valid
patter for `MATCH AGAINST` query
Test Plan:
call the endpoint with a strinng that had onnly stopwords, checked that there are noo errors. Called the endpoit with a string containning both a stopword and a non stopword, checked that it
returns correct values. Checked he endpoit with a string containning only a non stopword.

Reviewers:
kamil, bartek,
Subscribers: