[web] Confirm a user wants to exit if changes are made in the modal
Summary:
Based on the designs, if a member's role is changed (not saved, but just different to their current role) in the dropdown menu, we need to add a confirmation exit modal to warn them that they will
lose their changes by closing the role change modal.

https://linear.app/comm/issue/ENG-4039/confirm-a-user-wants-to-exit-if-changes-are-made-in-the-modal

Depends on D8138

Test Plan:
Please see the video below that covers the different use cases of no changes, changes, making a change but then changing the role back to the member's current role, etc.

{F575232}

Reviewers: atul, ginsu

Reviewed By: atul

Subscribers: ashoat, tomek, ted

Differential Revision: https://phab.comm.dev/D8139