[web] Add functionality on save to actually change a member's role
Summary:
This diff adds functionality to clicking the save button, so when it is clicked the member's role is actually changed. There is already keyserver support for this, so the work here was entirely front
end.

https://linear.app/comm/issue/ENG-4040/add-functionality-on-save-to-actually-change-a-members-role

Depends on D8139

Test Plan: Please see the video below showing the working save button

Reviewers: atul, ginsu

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8140